refactor(FileLocations): move to correct libs/folders

+++ BeestjeOpJeFeestje/BusinessLogic/Interfaces/IPricingRules.cs
@@ -4,6 +4,6 @@ namespace BusinessLogic.Interfaces {
     public interface IPricingRules {
 
         public double CalculateAnimalsPrice(List<Animal> selectedAnimals);
-        public (double, List<string>) CalculateDiscount(List<Animal> selectedAnimals, CustomerCard customerCard, DateTime bookingDate);
+        public (int, List<string>) CalculateDiscount(List<Animal> selectedAnimals, CustomerCard customerCard, DateTime bookingDate);
     }
 }
